% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_functions.R
\name{plot_htp}
\alias{plot_htp}
\title{Plot Multi-Temporal High-Throughput Phenotyping (HTP) Products}
\usage{
plot_htp(
  output_path = "PROCESING/STACKS_07_03_26V2/",
  multi_path = "PROCESING/MULTI/",
  target_cod = "CQC-183",
  target_bloq = "B1",
  border_file = "PROCESING/BORDERS/BORDERS_DAP_62_bloq.gpkg",
  daps = c("62", "86", "93", "121", "128"),
  plot_path = NULL
)
}
\arguments{
\item{output_path}{Character. Path to the directory containing processed index
rasters (e.g., CC, CV, NDVI, ExG folders).}

\item{multi_path}{Character. Path to the directory containing original
multispectral/RGB files.}

\item{target_cod}{Character. The unique identifier (\code{COD}) of the plot to visualize.}

\item{target_bloq}{Character. The blocking identifier (\code{BLOQ}) of the plot.}

\item{border_file}{Character. Path to the vector file (e.g., .gpkg) containing
plot boundaries.}

\item{daps}{A character vector of Days After Planting to include in the time series.}

\item{plot_path}{Character. Path where the final resulting image (.png) will be saved.}
}
\value{
A \code{patchwork} ggplot object. The function also saves a PNG file
to the specified \code{plot_path}.
}
\description{
This function generates a grid of plots for a specific experimental unit (plot),
showing its progression across multiple Days After Planting (DAP). It displays
RGB imagery alongside Canopy Cover, Canopy Volume, NDVI, and ExG indices.
}
\details{
The function crops all rasters to the extent of the selected plot and applies
a mask. It uses \code{patchwork} to arrange the indices as rows and DAPs as columns,
providing a clear visual overview of temporal vegetation dynamics.
}
